+++ src/app/src/components/ClaimedFacilities.jsx
@@ -4,6 +4,7 @@ import { Switch, Route } from 'react-router-dom';
 import ClaimedFacilitiesList from './ClaimedFacilitiesList';
 import ClaimedFacilitiesDetails from './ClaimedFacilitiesDetails';
 import RouteNotFound from './RouteNotFound';
+import AppOverflow from './AppOverflow';
 import AppGrid from './AppGrid';
 
 import {
@@ -13,30 +14,32 @@ import {
 
 export default function ClaimedFacilities() {
     return (
-        <AppGrid
-            title={
-                (
-                    <Switch>
-                        <Route
-                            path={claimedFacilitiesDetailRoute}
-                            render={() => 'Claimed Facility Details'}
-                        />
-                        <Route render={() => 'Claimed Facilities'} />
-                    </Switch>
-                )
-            }
-        >
-            <Switch>
-                <Route
-                    path={claimedFacilitiesDetailRoute}
-                    component={ClaimedFacilitiesDetails}
-                />
-                <Route
-                    path={claimedFacilitiesRoute}
-                    component={ClaimedFacilitiesList}
-                />
-                <Route render={() => <RouteNotFound />} />
-            </Switch>
-        </AppGrid>
+        <AppOverflow>
+            <AppGrid
+                title={
+                    (
+                        <Switch>
+                            <Route
+                                path={claimedFacilitiesDetailRoute}
+                                render={() => 'Claimed Facility Details'}
+                            />
+                            <Route render={() => 'Claimed Facilities'} />
+                        </Switch>
+                    )
+                }
+            >
+                <Switch>
+                    <Route
+                        path={claimedFacilitiesDetailRoute}
+                        render={() => <Route component={ClaimedFacilitiesDetails} />}
+                    />
+                    <Route
+                        path={claimedFacilitiesRoute}
+                        component={ClaimedFacilitiesList}
+                    />
+                    <Route render={() => <RouteNotFound />} />
+                </Switch>
+            </AppGrid>
+        </AppOverflow>
     );
 }